+++ pysmt/cmd/install.py
@@ -155,7 +155,7 @@ def download(url, file_name):
 def install_msat(options):
     """Installer for the MathSAT5 solver python interafce"""
 
-    base_name =  "mathsat-5.2.12-linux-%s" % get_architecture()
+    base_name =  "mathsat-5.3.4-linux-%s" % get_architecture()
     archive_name = "%s.tar.gz" % base_name
     archive = os.path.join(BASE_DIR, archive_name)
     dir_path = os.path.join(BASE_DIR, base_name)
@@ -415,7 +415,7 @@ def install_picosat(options):
 def check_install():
     """Checks which solvers are visible to pySMT."""
 
-    from pysmt.shortcuts import Solver
+    from pysmt.shortcuts import Solver, QuantifierEliminator
     from pysmt.exceptions import NoSolverAvailableError
 
     required_solver = os.environ.get("PYSMT_SOLVER")
@@ -425,7 +425,7 @@ def check_install():
         # Special case for bdd
         required_solver = "bdd"
 
-
+    print("Solvers:")
     for solver in ['msat', 'z3', 'cvc4', 'yices', 'bdd', 'picosat']:
         is_installed = False
         try:
@@ -433,7 +433,20 @@ def check_install():
             is_installed = True
         except NoSolverAvailableError:
             is_installed = False
-        print("%s%s" % (solver.ljust(10), is_installed))
+        print("  %s%s" % (solver.ljust(10), is_installed))
+
+        if solver == required_solver and not is_installed:
+            raise Exception("Was expecting to find %s installed" % required_solver)
+
+    print("\nQuantifier Eliminators:")
+    for solver in ['msat_fm', 'msat_lw', 'z3']:
+        is_installed = False
+        try:
+            QuantifierEliminator(name=solver)
+            is_installed = True
+        except NoSolverAvailableError:
+            is_installed = False
+        print("  %s%s" % (solver.ljust(10), is_installed))
 
         if solver == required_solver and not is_installed:
             raise Exception("Was expecting to find %s installed" % required_solver)