+++ pysmt/factory.py
@@ -33,7 +33,7 @@ from pysmt.logics import most_generic_logic, get_closer_logic
 from pysmt.oracles import get_logic
 
 DEFAULT_SOLVER_PREFERENCE_LIST = ['msat', 'z3', 'cvc4', 'yices', 'picosat', 'bdd']
-DEFAULT_QELIM_PREFERENCE_LIST = ['z3']
+DEFAULT_QELIM_PREFERENCE_LIST = ['z3', 'msat_fm', 'msat_lw']
 DEFAULT_LOGIC = QF_UFLIRA
 
 class Factory(object):
@@ -203,6 +203,14 @@ class Factory(object):
         try:
             from pysmt.solvers.z3 import Z3QuantifierEliminator
             self._all_qelims['z3'] = Z3QuantifierEliminator
+        except ImportError:
+            pass
+
+        try:
+            from pysmt.solvers.msat import (MSatFMQuantifierEliminator,
+                                            MSatLWQuantifierEliminator)
+            self._all_qelims['msat_fm'] = MSatFMQuantifierEliminator
+            self._all_qelims['msat_lw'] = MSatLWQuantifierEliminator
         except ImportError:
             pass